% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colours.R
\docType{data}
\name{Bede_colours}
\alias{Bede_colours}
\title{Colour Functions}
\format{
An object of class \code{character} of length 29.
}
\usage{
Bede_colours
}
\arguments{
\item{Colours}{used to create ggplot scales}
}
\description{
These function add colours to add to ggplot scales
}
\examples{
Bede_colours()
Bede_cols()
Bede_Palettes()
Bede_pal()
}
\keyword{Colours}
\keyword{Palettes}
