% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeFun.R
\name{completeFun}
\alias{completeFun}
\title{Complete Fun}
\usage{
completeFun(data, desiredCols)
}
\arguments{
\item{Remove}{NA rows}
}
\description{
This function remove NA rows
}
\examples{
completeFun()
}
\keyword{NAs}
\keyword{Remove}
