% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertRow.R
\name{insertRow}
\alias{insertRow}
\title{Insert Row}
\usage{
insertRow(existingDF, newrow, r)
}
\arguments{
\item{Insert}{rows}
}
\description{
This function adds a row to a df
}
\examples{
insertRow()
}
\keyword{Insert}
