% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Check.R
\name{Model_Check}
\alias{Model_Check}
\title{A Model Diagnostic Function Plot}
\usage{
Model_Check(Model)
}
\arguments{
\item{Model}{used to diagnose model assumptions}
}
\description{
This function Plots Model Diagnostics
}
\examples{
Model_Check()
}
\keyword{Diagnostics}
\keyword{Model}
