% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_Bede_Orkney.R
\name{plot_model_Bede_Orkney}
\alias{plot_model_Bede_Orkney}
\title{Model Plot Function}
\usage{
plot_model_Bede_Orkney(df_mean, df_pred, Metric, Formula)
}
\arguments{
\item{plot}{used to create ggplot scales}
}
\description{
This Function creates a ggplot of Model Predictions using YearNum and Area
}
\examples{
plot_model_Bede_Orkney()
}
\keyword{Model}
\keyword{Predictions}
