% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_Bede.R
\name{scale_colour_Bede}
\alias{scale_colour_Bede}
\title{Scale Colour Function}
\usage{
scale_colour_Bede(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{Colours}{used to create ggplot scales}
}
\description{
This function add colours to add to ggplot scales
}
\examples{
scale_colour_Bede()
}
\keyword{Colours}
\keyword{Palettes}
