% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_Clara.R
\name{scale_colour_Clara}
\alias{scale_colour_Clara}
\title{Scale Colour Function}
\usage{
scale_colour_Clara(palette = "PhD", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{Colours}{used to create ggplot scales}
}
\description{
This function add colours to add to ggplot scales
}
\examples{
scale_colour_Clara()
}
\keyword{Colours}
\keyword{Palettes}
