% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_Bede.R
\name{scale_fill_Bede}
\alias{scale_fill_Bede}
\title{Scale Fill Functions}
\usage{
scale_fill_Bede(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{fills}{used to create ggplot scales}
}
\description{
These function add colours to add to ggplot fill scales
}
\examples{
scale_fill_Bede()
}
\keyword{Fills}
\keyword{Palettes}
