% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard Error Function}
\usage{
se(x)
}
\arguments{
\item{Standard_Error}{function to create standard error}
}
\description{
This function calculates Standard Error
}
\examples{
se()
}
\keyword{se}
