% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Amelia.R
\name{theme_Amelia}
\alias{theme_Amelia}
\title{A ggplot Theme Function}
\usage{
theme_Amelia(...)
}
\arguments{
\item{Theme}{used to display ggplots}
}
\description{
This function adds a theme to ggplots
}
\examples{
theme_Amelia()
}
\keyword{Theme}
