% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Bede.R
\name{theme_Bede}
\alias{theme_Bede}
\title{A ggplot Theme Function}
\usage{
theme_Bede(...)
}
\arguments{
\item{Theme}{used to display ggplots}
}
\description{
This function adds a theme to ggplots
}
\examples{
theme_Bede()
}
\keyword{Theme}
