% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Bede_Storms.R
\name{theme_Bede_Storms}
\alias{theme_Bede_Storms}
\title{A ggplot Theme Function}
\usage{
theme_Bede_Storms(...)
}
\arguments{
\item{Theme}{used to display ggplots}
}
\description{
This function adds a theme to ggplots
}
\examples{
theme_Bede_Storms()
}
\keyword{Theme}
